+++ ReleaseNotes.md
@@ -1,3 +1,9 @@
+Version: 0.0.5
+Date: 2023.12.16
+Changes:
+- Optimized the removing of silence in transcription data
+- Mute the processing audio in parts where no singing is detected
+
 Version: 0.0.4
 Date: 2023.12.16
 Changes: